@@ -1,9 +1,28 @@
 package main
 
 import (
+	"context"
+	"github.com/aws/aws-sdk-go-v2/config"
+	"github.com/aws/aws-sdk-go-v2/service/route53"
 	"log"
+	"os"
 	"testing"
 )
+
+var (
+	ctx    context.Context
+	client *route53.Client
+)
+
+func TestMain(m *testing.M) {
+	ctx = context.Background()
+	cfg, err := config.LoadDefaultConfig(ctx)
+	if err != nil {
+		panic(err)
+	}
+	client = route53.NewFromConfig(cfg)
+	os.Exit(m.Run())
+}
 
 func TestGetCurrentIP(t *testing.T) {
 	ip, err := GetCurrentIP()
@@ -14,11 +33,11 @@ func TestGetCurrentIP(t *testing.T) {
 }
 
 func TestGetRecordIP(t *testing.T) {
-	id, err := GetZoneID("h.2k0ri.org")
+	id, err := GetZoneID(ctx, client, "h.2k0ri.org")
 	if err != nil {
 		t.Fail()
 	}
-	ip, err := GetRecordIP("h.2k0ri.org", id)
+	ip, err := GetRecordIP(ctx, client, "h.2k0ri.org", id)
 	if err != nil {
 		t.Fail()
 	}
@@ -26,7 +45,7 @@ func TestGetRecordIP(t *testing.T) {
 }
 
 func TestGetZoneID(t *testing.T) {
-	id, err := GetZoneID("h.2k0ri.org")
+	id, err := GetZoneID(ctx, client, "h.2k0ri.org")
 	if err != nil {
 		t.Fail()
 	}
